pyproject.toml
[build-system]
requires = ["uv-build>=0.6.6,<0.7"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest~=8.3.5",
    "ruff~=0.11.5",
]

[project]
name = "pre-commit-image"
version = "0.0.0"
description = "A pre-commit hook to process images"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "Subin Kim", email = "sudosubin@gmail.com" },
]
classifiers = [
    "Development Status :: 1 - Planning",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
]
urls = { "Repository" = "https://github.com/sudosubin/pre-commit-image" }
dependencies = []

[project.optional-dependencies]
avif = ["pillow-avif-plugin>=1.4.6,<2.0"]
heif = ["pillow-heif>=0.18.0,<0.23.0"]
svg = ["scour>=0.38,<0.39"]

[project.scripts]
pre-commit-image = "pre_commit_image.__main__:main"

[tool.ruff]
line-length = 80
src = ["src"]
target-version = "py38"

[tool.ruff.format]
preview = true

[tool.ruff.lint]
preview = true
select = ["ANN", "B", "E", "F", "FURB", "I", "ICN", "N", "RUF", "SIM", "TID"]

[tool.ruff.lint.flake8-import-conventions]
banned-from = ["typing"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
combine-as-imports = true
